Move the app folder into dokum folder
This commit takes action on the move application files, particularly, those related to app configuration, views, and controllers. It includes the movement of templates, partials, and scaffolds. Also, default controller files and configuration files used for setting up the application are moved. The intention behind moving these files serves as preparation for a complete overhaul of the application structure.